Fix address reading and user int type

+++ src/common/command/getaccount.cpp
@@ -8,7 +8,7 @@ GetAccount::GetAccount()
     m_responseError = ErrorCodes::Code::eNone;
 }
 
-GetAccount::GetAccount(uint16_t abank, uint32_t auser, uint16_t bbank, uint16_t buser,
+GetAccount::GetAccount(uint16_t abank, uint32_t auser, uint16_t bbank, uint32_t buser,
                        uint32_t time)
     : m_data( abank, auser, bbank, buser, time) {
     m_responseError = ErrorCodes::Code::eNone;
